+++ src/lib.rs
@@ -11,7 +11,7 @@
 //!     .set_prompt("!:") // set the prompt
 //!     .enable_alternate_screen() // use alternate screen
 //!     .disable_ctrl_c() // pressing control+c won't terminate the caprice console
-//!     .set_keywords(&[
+//!     .set_keywords(vec![
 //!         "some_token".to_owned(),
 //!         "some_other_token".to_owned(),
 //!         "exit".to_owned(), // an exit keyword